moved most of the fixtures to conftest
